# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "basePath": "/apis",
    "paths": {
        "/liveness": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Misc"
                ],
                "summary": "K8s Liveness endpoint",
                "responses": {
                    "204": {
                        "description": "No Content"
                    }
                }
            }
        },
        "/readiness": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Misc"
                ],
                "summary": "K8s Readiness endpoint",
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/ResponseError"
                        }
                    }
                }
            }
        },
        "/version": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Misc"
                ],
                "summary": "Get version of GPT X-Plane plugin",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ResponseOk"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/ResponseError"
                        }
                    }
                }
            }
        },
        "/xplm/dataref": {
            "get": {
                "security": [
                    {
                        "Oauth2Application": [
                            ""
                        ]
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Dataref"
                ],
                "summary": "Get Dataref",
                "parameters": [
                    {
                        "type": "string",
                        "description": "xplane dataref string",
                        "name": "dataref_str",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "alias name, if not set, dataref_str will be used",
                        "name": "alias",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "-1: raw, 2: round up to two digits",
                        "name": "precision",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "boolean",
                        "description": "transform xplane byte array to string",
                        "name": "is_byte_array",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.DatarefValue"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/ResponseError"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "Oauth2Application": [
                            ""
                        ]
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Dataref"
                ],
                "summary": "Set Dataref",
                "parameters": [
                    {
                        "description": "dataref and value",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.SetDatarefValue"
                        }
                    }
                ],
                "responses": {
                    "501": {
                        "description": "Not Implemented"
                    }
                }
            }
        },
        "/xplm/datarefs": {
            "put": {
                "security": [
                    {
                        "Oauth2Application": [
                            ""
                        ]
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Dataref"
                ],
                "summary": "Set a list of Dataref",
                "responses": {
                    "501": {
                        "description": "Not Implemented"
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "Oauth2Application": [
                            ""
                        ]
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Dataref"
                ],
                "summary": "Get a list of Dataref",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.DatarefValue"
                            }
                        }
                    },
                    "501": {
                        "description": "Not Implemented"
                    }
                }
            }
        }
    },
    "definitions": {
        "ResponseError": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
        "ResponseOk": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
        "models.DatarefValue": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                },
                "value": {}
            }
        },
        "models.SetDatarefValue": {
            "type": "object",
            "properties": {
                "dataref": {
                    "type": "string"
                },
                "value": {}
            }
        }
    },
    "securityDefinitions": {
        "": {
            "type": "oauth2",
            "flow": "accessCode",
            "authorizationUrl": "https://pluginlab.xairline.org/oauth/authorize",
            "tokenUrl": "https://pluginlab.xairline.org/oauth/token"
        }
    }
}